Update timetable_creator.html
I switched the timetable conflict warning. A growing voice inside of me says that I should only ever commit something that I think could somehow slip into production. Left it initially as a 'this will remind me to come back and redo it', but I am seeing now that that attitude is a lame attitude.

The conflict warning is something I want to discuss though, and right now it is awfully obnoxious.