Modals: Tabs now hidden when modal is displayed (shown when modal is closed).
I think that the tabs need to be superfically changed,
as well as disabled.
The user should understand that the tabs won't work when modals are displayed.
This can either be done by hiding them, or doing something like greying them out (much like Ubuntu does when a process is not responding).

I think that the former is better and cleaner.