database: Insertion now works.
Perhaps this is not the best way to perform insertion.
I would rather perform all SQL actions within one larger call
of runSqlite, but I have been having difficulty with performing
actions with ReaderT Sqlbackend MonadIOs (or such) within an IO ()
function. However, this does perform the action that we want to perform,
but does not do it in :memory:. Rather, it performs all actions from a
file in the directory of execution.