database: Preliminary Tutorial insertion.
Tutorials are formatted as [Text, [[Int]], Text]. Tutorials need to be
parsed in order to maintain a healthy database, however for the
purposes of the objective, which is to reconstruct JSON strings from
database values, I have decided to leave it as a text blob.

In future commits, I plan to properly parse the Tutorial.