tutorial-parse-fix: Fix tutorial mix up. Read more...
In the JSON files, some tutorials are represented by an array with 3 values.
E.g. CSC148H1.txt. The fromJSON would only parse an array of length 2.
This would work most of the time, but would fail for tutorials that
included the third bit, the tutorial section id.

This has been fixed in the Tutorial fromJSON declaration.