react-grid: Add partial commit of course closing.
The courses of the grid currently doesn't close properly. To demonstrate this, click the red "x" of the first course at index 0 after loading the page. The state of course that was previously after it, at index 1, is now true, though this course is now at index 0 instead of 1 because the course at index 0 was removed. Now click the red "x" again, the course at index 1, which is now at index 0, has no change in the state. Which shows that the course at index "y" after having the red "x" clicked, changes the "satisifed" state of the course at index "y + 1" to the opposite of the state at index "y."