Button to deselect course in Sidebar (#1413)
* Add an X button next to each select course in the sidebar panel to deselect a course.

In the panel list of selected courses, currently the only way to deselect is by clicking again on the course node or at the course code link within the panel (this has the FCE bug described in #1). The latter method is counter-intuitive and undesired. Instead, we want to add an **X** button next to each course code in the panel so that a given course gets deselected when the user clicks on that button (and not when they click on the course code link). The required changes in the Sidebar and Graph components.

* clicking sidebar course-link opens course info modal