Improved style
(1) disabled dependency highlighting in focus-mode to avoid conflict, (2) Remove any kind of coloured background when hughlighting dependencies, as it leads to too much traffic, (3) Have a different colored spotlight for focus-highlighting vs dependency-highlighting, (4) During dependency highlighting, avoid coloring the node rectangle to avoid traffic. Perhaps the most important change is: Now, when hovering over a node, there is no color indicating whether that course is a missing or taken. However, this does not matter because highlighting completely ignores taken nodes and their ancestors, so the only case that differs only the currently hovered node. Of course the feature still exists for when clicking different nodes.