Added circling for course dependencies on hover (#1437)
* Added spotlight for course dependencies on hover
* Fixed Bug in Sidebar & Applied same circling functional for the sidebar too

When searching on the sidebar, courses appear, and the intended behavior of clicking a course would be to add it to the state-list of selected courses, but instead now the course info was opening.

Remove any kind of coloured background when highlighting dependencies, as it leads to too much visual noise, Now, when hovering over a node, there is no color indicating whether that course is a missing or taken. However, this does not matter because highlighting completely ignores taken nodes and their ancestors, so the only case that differs only the currently hovered node. Of course the feature still exists for when clicking different nodes.

Render a different colored spotlight for focus-highlighting vs dependency-highlighting, 

Separated logic between highlighting focus and highlighting dependencies This independence allows a more clean coexistence in the case of the focus tab, and will allow them to grow individually in the future (instead of fearing that change in one will break the other)

Made ellipse proportionally bigger for combo courses -- e.g. MAT and STAT course blocks.